# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TallerFlow API
  version: 0.1.0
  description: API contract for TallerFlow Phase 1 secure foundations.
servers:
  - url: /
    description: Same-origin deployment
...
paths:
...
  /api/v1/auth/change-password:
    post:
      tags: [Authentication]
      operationId: changePassword
      summary: Change the current user's password and rotate the session
      description: >-
        Revokes the user's other sessions, clears the mandatory-change flag,
        rotates the current opaque session, and returns the replacement CSRF
        token. Password confirmation is a client-side concern and is not sent.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordInput'
      responses:
        '200':
          description: Password changed and current session rotated.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Set-Cookie:
              $ref: '#/components/headers/SessionCookieSet'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/SessionInvalid'
        '403':
          $ref: '#/components/responses/CSRFInvalid'
        '422':
          description: Current password is wrong or the new password violates policy.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-tallerflow_session
      description: >-
        Opaque session token. The server sets this cookie with Secure, HttpOnly,
        SameSite=Lax, and Path=/; browser code must not read or persist it.
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: >-
        Raw token bound to the current session. Required on every authenticated
        mutation and validated together with Origin or Referer.
      schema:
        type: string
        minLength: 43
        maxLength: 43
  headers:
    RequestID:
      description: Unique identifier assigned to the request.
      schema:
        type: string
        format: uuid
    SessionCookieSet:
      description: >-
        Sets __Host-tallerflow_session with Secure, HttpOnly, SameSite=Lax, and
        Path=/. Its value is an opaque token and is never returned in JSON.
      schema:
        type: string
        example: __Host-tallerflow_session=opaque-token; Path=/; Secure; HttpOnly; SameSite=Lax
    SessionCookieCleared:
      description: Expires the opaque session cookie immediately.
      schema:
        type: string
        example: __Host-tallerflow_session=; Path=/; Max-Age=0; Secure; HttpOnly; SameSite=Lax
  responses:
    InvalidRequest:
      description: Request body or parameters are invalid.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    SessionInvalid:
      description: Session is missing, expired, revoked, or otherwise invalid.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
        Set-Cookie:
          $ref: '#/components/headers/SessionCookieCleared'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidSession:
              value:
                error:
                  code: SESSION_INVALID
                  message: The session is invalid or has expired.
                  details: {}
                  request_id: 3c0d522d-ec6d-4d83-aac5-40fef2e2c004
    CSRFInvalid:
      description: CSRF token or request origin validation failed.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidCSRF:
              value:
                error:
                  code: CSRF_INVALID
                  message: CSRF validation failed.
                  details: {}
                  request_id: 3c0d522d-ec6d-4d83-aac5-40fef2e2c004
    InternalError:
      description: Unexpected server error.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ChangePasswordInput:
      type: object
      additionalProperties: false
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          minLength: 1
          maxLength: 1048576
          writeOnly: true
        new_password:
          type: string
          minLength: 12
          maxLength: 1048576
          writeOnly: true
          description: Must differ from the current password.
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, details, request_id]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          example: SESSION_INVALID
        message:
          type: string
          minLength: 1
          example: The session is invalid or has expired.
        details:
          type: object
          additionalProperties: true
          description: Safe, machine-readable context; never contains secrets.
        request_id:
          type: string
          format: uuid
    SessionData:
      type: object
      additionalProperties: false
      required: [expires_at, csrf_token, must_change_password]
      properties:
        expires_at:
          type: string
          format: date-time
          description: UTC expiry of the current session, at most eight hours after creation.
        csrf_token:
          type: string
          minLength: 43
          maxLength: 43
          description: >-
            Raw CSRF token for this session. Keep only in memory and send it in
            X-CSRF-Token on authenticated mutations.
        must_change_password:
          type: boolean
          description: >-
            When true, only session restoration, password change, and logout
            flows are available until the password is replaced.
...
    Meta:
      type: object
      additionalProperties: true
      description: Optional response metadata.
...
    SessionResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/SessionData'
        meta:
          $ref: '#/components/schemas/Meta'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'